# backend/src/app/src/openapi_server/models/closest_winners.py
# coding: utf-8

"""
    Donator-Mainframe communication

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 0.1.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictStr, field_validator
from typing import Any, ClassVar, Dict, List
from openapi_server.models.abgeordneter import Abgeordneter
from openapi_server.models.partei import Partei
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class ClosestWinners(BaseModel):
    """
    ClosestWinners
    """ # noqa: E501
    party: Partei
    closest_type: StrictStr
    closest_winners: List[Abgeordneter] = Field(alias="closestWinners")
    __properties: ClassVar[List[str]] = ["party", "closest_type", "closestWinners"]

    @field_validator('closest_type')
    def closest_type_validate_enum(cls, value):
        """Validates the enum"""
        if value not in ('Winner', 'Loser',):
            raise ValueError("must be one of enum values ('Winner', 'Loser')")
        return value

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of ClosestWinners from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of party
        if self.party:
            _dict['party'] = self.party.to_dict()
        # override the default output from pydantic by calling `to_dict()` of each item in closest_winners (list)
        _items = []
        if self.closest_winners:
            for _item in self.closest_winners:
                if _item:
                    _items.append(_item.to_dict())
            _dict['closestWinners'] = _items
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of ClosestWinners from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "party": Partei.from_dict(obj.get("party")) if obj.get("party") is not None else None,
            "closest_type": obj.get("closest_type"),
            "closestWinners": [Abgeordneter.from_dict(_item) for _item in obj.get("closestWinners")] if obj.get("closestWinners") is not None else None
        })
        return _obj